backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gunicorn>=23.0.0
python-multipart>=0.0.9
pydantic-settings>=2.5.0
python-dotenv>=1.0.1

# Azure AI SDKs
azure-cognitiveservices-vision-customvision>=3.1.0
azure-ai-vision-imageanalysis>=1.0.0
azure-ai-contentsafety>=1.0.0
azure-ai-textanalytics>=5.3.0
openai>=1.45.0

# Azure Storage
azure-storage-blob>=12.22.0

# Database
aiosqlite>=0.20.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
google-auth>=2.29.0
email-validator>=2.0.0

# Image Processing
Pillow>=10.4.0

# HTTP Client
httpx>=0.27.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
ruff>=0.7.0
